NGEN-57 Refactor old generation core
