+++ src/app/generation/core/algorithm-segments/name-ending-applier.ts
@@ -1,7 +1,8 @@
-import { LetterUtils, RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
 import { LetterSet, RegularString } from "@ngen-shared/models";
 import { RandomUtils } from "@ngen-shared/utils";
-import { AlgorithmDataType, AlgorithmSegment, GenerationData } from "../models";
+import { AlgorithmDataType } from "../enums";
+import { AlgorithmSegment, GenerationData } from "../models";
+import { LetterUtils, RegularUtils } from "../utils";
 
 export interface NameEndingApplierConfig {
    excludedLetters: LetterSet;
@@ -11,7 +12,7 @@ export interface NameEndingApplierConfig {
 export class NameEndingApplier extends AlgorithmSegment<
    AlgorithmDataType.REGULAR_STRING,
    AlgorithmDataType.REGULAR_STRING,
-   {}
+   NameEndingApplierConfig
 > {
    private static readonly DOUBLE_CONSONANT_ENDINGS = ["lf", "lv", "lt", "ld", "mn", "rk", "ck"];
    private static readonly NAME_ENDINGS = [
