+++ src/app/generation/core/algorithm-segments/template-based-regular-structure-generator.ts
@@ -1,7 +1,8 @@
-import { RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
 import { RegularString } from "@ngen-shared/models";
 import { RandomUtils } from "@ngen-shared/utils";
-import { AlgorithmDataType, AlgorithmSegment, GenerationData } from "../models";
+import { AlgorithmDataType } from "../enums";
+import { AlgorithmSegment, GenerationData } from "../models";
+import { RegularUtils } from "../utils";
 
 export interface TemplateBasedRegularStructureGeneratorConfig {
    minLength: number;
