+++ src/app/generation/core/constants/index.ts
@@ -1,0 +1,3 @@
+//@index("./*.ts", (f, _) => `export { ${_.camelCase(f.name)} } from "${f.path}";`)
+export { japaneseLetterList } from "./japanese-letter-list";
+//@endindex
