+++ src/app/generation/core/enums/algorithm-data-type.ts
