+++ src/app/generation/core/enums/generator-algorithm-name.ts
