+++ src/app/generation/core/enums/index.ts
@@ -1,3 +1,4 @@
 //@index("./*.ts", (f, _) => `export { ${_.pascalCase(f.name)} } from "${f.path}";`)
+export { AlgorithmDataType } from "./algorithm-data-type";
 export { GeneratorAlgorithmName } from "./generator-algorithm-name";
 //@endindex
