+++ src/app/generation/core/models/generation-config.d.ts
