+++ src/app/generation/core/models/index.ts
@@ -1,6 +1,9 @@
 //@index("./*", (f, _) => `export { ${_.pascalCase(f.name.split(".")[0])} } from "${f.path}";`)
-export { AlgorithmDataType } from "./algorithm-data-type";
 export { AlgorithmSegment } from "./algorithm-segment";
+export { GenerationConfig } from "./generation-config.d";
 export { GenerationData } from "./generation-data.d";
+export { JapaneseLetter } from "./japanese-letter.d";
+export { Letter } from "./letter.d";
 export { NameGeneratorAlgorithm } from "./name-generator-algorithm";
+export { RandomLetterConfig } from "./random-letter-config.d";
 //@endindex
