+++ src/app/generation/core/models/japanese-letter.d.ts
