+++ src/app/generation/core/models/letter.d.ts
