+++ src/app/generation/core/models/random-letter-config.d.ts
