+++ src/app/generation/core/utils/letter.utils.ts
