+++ src/app/generation/core/utils/regular.utils.spec.ts
