+++ src/app/generation/core/utils/regular.utils.ts
