--- src/app/generation/generator-algorithms/generator-service.model.ts
@@ -1,6 +1,0 @@
-import { GenerationConfig } from "@ngen-generation/models";
-import { Name } from "@ngen-shared/names";
-
-export interface GeneratorService {
-   generateName(config: GenerationConfig): Name;
-}
