--- src/app/generation/generator-algorithms/index.ts
@@ -1,3 +1,0 @@
-export { JapaneseGeneratorService } from "./japanese/japanese-generator.service";
-export { RegularGeneratorService } from "./regular/regular-generator.service";
-export { SyllabicGeneratorService } from "./syllabic/syllabic-generator.service";
