--- src/app/generation/generator-algorithms/letter-finalization/models/index.ts
@@ -1,4 +1,0 @@
-//@index("./*", (f, _) => `export { ${_.pascalCase(f.name)} } from "${f.path}";`)
-export { Letter } from "./letter";
-export { RandomLetterConfig } from "./random-letter-config";
-//@endindex
