--- src/app/generation/generator-algorithms/name-endings/index.ts
@@ -1,4 +1,0 @@
-//@index("./*", (f, _) => `export { ${ _.camelCase(f.name) } } from "${f.path}";`)
-export { matchNameEnding } from "./match-name-ending";
-export { nameEndings } from "./name-endings";
-//@endindex
