+++ src/app/generation/models/index.ts
@@ -1,4 +1,3 @@
 //@index("./*", (f, _) => `export { ${_.pascalCase(f.name)} } from "${f.path}";`)
-export { GenerationConfig } from "./generation-config";
 export { GeneratorFlagKey } from "./generator-flag-key";
 //@endindex
