+++ src/app/generation/services/configuration-store.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from "@angular/core";
-import { GeneratorAlgorithmName } from "@ngen-generation/enums";
+import { GeneratorAlgorithmName } from "@ngen-generation/core/enums";
+import { GenerationConfig } from "@ngen-generation/core/models";
 import {
    japaneseDefaultConfig,
    regularDefaultConfig,
    syllabicDefaultConfig
 } from "@ngen-generation/generation-config/default-configs";
-import { GenerationConfig } from "@ngen-generation/models";
 
 @Injectable({
    providedIn: "root"
