+++ src/app/shared/models/regular-character.ts
@@ -1,4 +1,4 @@
-import { LetterUtils, RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
+import { LetterUtils, RegularUtils } from "@ngen-generation/core/utils";
 import { RegularCharacterPriority } from "@ngen-shared/enums";
 
 enum RegularCharacterType {
