+++ src/app/shared/models/regular-reference.ts
@@ -1,4 +1,4 @@
-import { RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
+import { RegularUtils } from "@ngen-generation/core/utils";
 import { RegularCharacterPriority } from "@ngen-shared/enums";
 import { RegularCharacter } from "./regular-character";
 import { RegularString } from "./regular-string";
